requirements.txt
openai>=1.3.0
anthropic>=0.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-nmap>=0.7.1
scapy>=2.5.0
paramiko>=3.3.1
pycryptodome>=3.19.0
colorama>=0.4.6
rich>=13.7.0
click>=8.1.7
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.9.0
selenium>=4.15.0
webdriver-manager>=4.0.1
sqlparse>=0.4.4
dnspython>=2.4.2
python-whois>=0.8.0
shodan>=1.30.1
censys>=2.2.8
